Fix atom type consistency and surface hotkey conflicts from Settings
- GlobalDeleteAtom P/Invoke parameter corrected to ushort (was int)
- Atom dictionaries and HandleHotkey now use ushort end-to-end,
  matching the Win32 ATOM type through to WndProc
- Hotkey conflicts that occur while editing profiles in the Settings
  window now show a balloon notification instead of being silently dropped

+++ VibeSwitcher/App.xaml.cs
@@ -95,7 +95,7 @@ public partial class App : Application
     {
         if (msg == WinApi.WM_HOTKEY)
         {
-            int atomId = wParam.ToInt32();
+            ushort atomId = (ushort)wParam.ToInt32();
             Guid profileId = _hotkeyService!.HandleHotkey(atomId);
             if (profileId != Guid.Empty)
             {
